add to current day, no testing

+++ back/src/main/java/com/EatingTracker/back/models/EatenItemModel.java
@@ -1,12 +1,20 @@
 package com.EatingTracker.back.models;
 
+import com.EatingTracker.back.entities.EatenItemType;
+import com.EatingTracker.back.entities.EatenListItem;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.parameters.P;
+
+import java.util.ArrayList;
+import java.util.List;
+
 public class EatenItemModel {
     private DisplayListItemModel item;
 
-    private int amm;
+    private float amm;
 
     public EatenItemModel(){}
-    public EatenItemModel(DisplayListItemModel item, int amm) {
+    public EatenItemModel(DisplayListItemModel item, float amm) {
         this.item = item;
         this.amm = amm;
     }
@@ -19,12 +27,13 @@ public class EatenItemModel {
         this.item = item;
     }
 
-    public int getAmm() {
+    public float getAmm() {
         return amm;
     }
 
-    public void setAmm(int amm) {
+    public void setAmm(float amm) {
         this.amm = amm;
     }
+
 }
